requirements.txt
asgiref==3.7.2
Django==4.2.8
django-js-asset==2.2.0
django-mptt==0.15.0
Pillow==9.5.0
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
